feat(phase-6): PostgreSQL + Redis job queue infrastructure
- Add multi-backend persistence layer (SQLite / PostgreSQL)
  - db_migrations.py: idempotent, versioned schema migrations
  - db.py: DatabaseConfig, protocol bases, factory with fallback
  - job_sqlite.py / user_sqlite.py: updated SQLite stores
  - job_postgres.py / user_postgres.py: new Postgres stores

- Add distributed Redis job queue with reliable execution
  - queue_service.py: RedisJobQueue (atomic ops, visibility timeout,
    exponential-backoff retries, dead-letter, stale recovery, heartbeat)
    and ThreadPoolJobQueue (local in-process fallback)

- Add standalone worker daemon (worker.py)
  - SIGTERM/SIGINT graceful shutdown
  - Periodic stale-job recovery sweeps

- Integrate diagnostics into health API (main.py)
  - GET /api/system/database  -- db backend status + migration version
  - GET /api/system/queue     -- queue backend + pending count
  - GET /health               -- enriched with database + queue sub-objects

- Infrastructure
  - docker-compose.yml / docker-compose.gpu.yml: postgres + redis services
  - requirements.txt: psycopg[binary], redis
  - .env.example / .env.production.example: new env vars

- Tests (397 passing, 0 failures)
  - test_db_persistence.py: migrations, SQLiteJobStore, SQLiteUserStore,
    DatabaseConfig, store factory + Postgres fallback
  - test_redis_queue.py: QueueConfig, ThreadPoolJobQueue, RedisJobQueue
    (enqueue, ack, fail/retry, dead-letter, recovery, heartbeat, diagnostics)
  - test_system_diagnostics_api.py: /health, /api/system/database,
    /api/system/queue endpoint contracts
  - Updated test_jobs_api.py + test_e2e_production_workflow.py to accept
    enriched /health response body

- Fix redundant int()/float() casts in service layer (clip_planner,
  dynamic_caption, karaoke_caption, smart_framing services)

+++ backend/app/services/dynamic_caption_service.py
@@ -214,7 +214,7 @@ class DynamicCaptionService:
             if i == len(chunks) - 1:
                 cur_end = segment.end_seconds
             else:
-                cur_end = float(round(cur_start + ch_duration, 3))
+                cur_end = round(cur_start + ch_duration, 3)
 
             # Ensure valid bounds
             cur_end = max(cur_start + 0.05, cur_end)
